CPM-807: Move back to identifier generator context
