+++ components/identifier-generator/back/src/Infrastructure/Event/UnableToSetIdentifierEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Automation\IdentifierGenerator\API\Event;
+namespace Akeneo\Pim\Automation\IdentifierGenerator\Infrastructure\Event;
 
 use Akeneo\Pim\Automation\IdentifierGenerator\Application\Exception\UnableToSetIdentifierException;
